+++ lib/datagrid/index.ts
@@ -1,0 +1,19 @@
+import {
+	Datagrid,
+	DatagridColumn,
+	DatagridProps,
+	DatagridRow,
+	DatagridState,
+	getDefaultDatagridProps,
+	getDefaultDatagridState
+} from "./Datagrid";
+
+export {
+	Datagrid,
+	DatagridColumn,
+	DatagridProps,
+	DatagridRow,
+	DatagridState,
+	getDefaultDatagridProps,
+	getDefaultDatagridState
+};